chore: refactor leaderboard components #110
