+++ index.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html>
+<html style="background-color: #f2f3f4">
   <head>
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
@@ -8,7 +8,7 @@
       content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no "
     />
     <link rel="favicon icon" href="arweave.svg" />
-    <title>arweave swag game</title>
+    <title>arweave swag stamper</title>
     <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto+Mono" />
     <script src="https://unpkg.com/arweave@1.13.0/bundles/web.bundle.min.js"></script>
     <script src="https://unpkg.com/warp-contracts@1.2.52/bundles/web.iife.bundle.min.js"></script>
