+++ src/components/leaderboard/row.svelte
